feat: userRoles

+++ src/services/practices.service.ts
@@ -1,37 +1,44 @@
-import { practiceRepository } from "@/repositories/practices.repository";
+import { unauthorized } from '@/errors/customErrors';
+import { UserCredentials } from '@/protocols';
+import { practiceRepository } from '@/repositories';
 
-export async function createPractice(name: string) {
-    return await practiceRepository.createPractice({ name });
+export async function createPractice(name: string, user: UserCredentials) {
+  if (user.role === 'ADMIN') throw unauthorized('Usuario não tem permissão');
+
+  return await practiceRepository.createPractice({ name });
 }
 
 export async function getPractice() {
-    return await practiceRepository.getPractice();
+  return await practiceRepository.getPractice();
 }
 
-export async function updatePractice(id: number, name: string) {
-    return await practiceRepository.updatePractice({ id, name });
+export async function updatePractice(id: number, name: string, user: UserCredentials) {
+  if (user.role === 'ADMIN') throw unauthorized('Usuario não tem permissão');
+
+  return await practiceRepository.updatePractice({ id, name });
 }
 
 export async function upsertPracticeAdvantage(
-    advantage: string,
-    description: string,
-    practiceId: number
+  advantage: string,
+  description: string,
+  practiceId: number,
+  user: UserCredentials,
 ) {
-    return await practiceRepository.upsertPracticeAdvantage(
-        advantage,
-        description,
-        practiceId
-    );
+  if (user.role === 'ADMIN') throw unauthorized('Usuario não tem permissão');
+
+  return await practiceRepository.upsertPracticeAdvantage(advantage, description, practiceId);
 }
 
-export async function deletePractice(practiceId: number) {
-    return await practiceRepository.deletePractice(practiceId);
+export async function deletePractice(practiceId: number, user: UserCredentials) {
+  if (user.role === 'ADMIN') throw unauthorized('Usuario não tem permissão');
+
+  return await practiceRepository.deletePractice(practiceId);
 }
 
 export const practiceServices = {
-    createPractice,
-    getPractice,
-    updatePractice,
-    upsertPracticeAdvantage,
-    deletePractice,
+  createPractice,
+  getPractice,
+  updatePractice,
+  upsertPracticeAdvantage,
+  deletePractice,
 };
